Split DevicesController: extract metadata mixin
Move _resolve_device_metadata, _derive_board_id_from_yaml,
_persist_device_ip_async, and _persist_device_metadata_async
into a sibling DeviceMetadataMixin in metadata.py. The
controller inherits the mixin (class DevicesController(
DeviceMetadataMixin)) so the methods become real attributes
on the instance via inheritance.

Departs from the free-function-with-controller-arg pattern
used in archive / firmware_sync / importable / logs /
reachability / storage_regen / api_key / add_component.
Mixin works cleanly here because the four methods only touch
self._db (the DeviceBuilder injection point) and same-mixin
siblings; no reach into _scanner / _state_monitor /
_reachability / _devices_by_name. The other modules access
controller-owned state heavily, so they justify the
explicit controller= arg threading; this one doesn't.

Test patches that targeted controller.set_device_metadata
or controller.get_device_metadata in code paths that moved
are re-pointed at metadata; tests that call the methods
as bound methods (controller._resolve_device_metadata etc.)
work unchanged because the mixin makes them bound methods.

Pure refactor; behaviour preserved. Drops ~134 lines from
controllers/devices/controller.py.
