+++ esphome_device_builder/controllers/devices/__init__.py
@@ -27,6 +27,13 @@ resolving after the subpackage split. Submodules:
   (``stream_logs``, ``stop_stream``) plus the shared
   ``stream_subprocess`` helper that ``validate_config``
   reuses.
+- ``metadata`` — ``DeviceMetadataMixin`` carrying
+  ``_resolve_device_metadata`` /
+  ``_derive_board_id_from_yaml`` /
+  ``_persist_device_ip_async`` /
+  ``_persist_device_metadata_async``. Mixed in directly
+  on ``DevicesController`` since the methods only touch
+  ``self._db`` and same-mixin siblings.
 - ``reachability`` — per-device reachability streaming + the
   on-subscription mDNS A-record refresh loop.
 - ``storage_regen`` — background ``--only-generate`` scheduler
